+++ crates/test-utils/src/lib.rs
@@ -537,7 +537,13 @@ impl Test {
     /// Returns the last state of the stack after executing a test.
     #[track_caller]
     pub fn get_last_stack_state(&self) -> StackOutputs {
-        let trace = self.execute().expect("failed to execute");
+        let trace = self
+            .execute()
+            .inspect_err(|_err| {
+                #[cfg(feature = "std")]
+                std::eprintln!("{}", PrintDiagnostic::new_without_color(_err))
+            })
+            .expect("failed to execute");
 
         trace.last_stack_state()
     }