+++ libcore/tests/crypto/mod.rs
@@ -5,6 +5,8 @@ mod aead;
 mod blake3;
 mod circuit_evaluation;
 mod ecdsa_k256_keccak;
+mod eddsa_ed25519;
 mod keccak256;
 mod rpo;
 mod sha256;
+mod sha512;